requirements.txt
tweepy==4.14.0
python-telegram-bot==20.7
requests==2.31.0